Phase 7b Parts 2+3: optional prefix keymap, arrow nav, hover-focus
Keymap:
- Keymap.prefix: KeyEvent → Option<KeyEvent> (breaking: callers wrap in Some)
- Keymap::new takes Option<KeyEvent> prefix
- Keymap::vim_prefix() — the old Ctrl-w two-key scheme, now explicit
- Keymap::default() — new prefixless scheme: plain arrows → FocusDir,
  Shift+arrows → FocusDirCross, Enter → ZoomIn, Esc → ZoomOut
- Keymap::prefix is now pub for introspection

NavCommand:
- FocusDir(Direction) — executed by router (calls tree.focus_dir)
- FocusDirCross(Direction) — returned as Nav outcome; app must call
  tree.focus_dir_cross(dir, area) since the router has no viewport

InputRouter:
- handle() branches on prefix.is_none() for flat lookup (no Consumed)
- hover_focus: bool field (default false)
- set_hover_focus(bool) opt-in; MouseMoved → Focused(id) when enabled
- examples updated to use InputRouter::with_keymap(Keymap::vim_prefix())

Also fix 6 pre-existing clippy lints in label.rs, layout.rs, tree.rs
(repeat_n, unnecessary u32 cast, index-only loop var, collapsible if,
map_or → is_some_and).

+++ src/layout.rs
@@ -358,9 +358,9 @@ pub(crate) fn solve_into(node: &mut Node, area: Rect, out: &mut Vec<(TileId, Rec
 
             for (child_idx, v_start, ext) in entries {
                 // Clip the child's content span to the viewport.
-                let vis_content_start = (v_start as u32).max(clamped as u32);
+                let vis_content_start = v_start.max(clamped as u32);
                 let vis_content_end =
-                    (v_start as u32 + ext as u32).min(clamped as u32 + main_extent as u32);
+                    (v_start + ext as u32).min(clamped as u32 + main_extent as u32);
                 let vis_len = (vis_content_end - vis_content_start) as u16;
                 // Screen offset from viewport origin for the visible portion.
                 let screen_start = vp_main_origin + (vis_content_start - clamped as u32) as u16;
